requirements.txt
google-genai>=1.0.0
tavily-python>=0.5.0
tweepy>=4.14.0
python-dotenv>=1.0.0
feedparser>=6.0.0
